Add ProgressBar component to poly-look
