+++ feature-utils/poly-look/src/react-components/index.js
@@ -35,4 +35,5 @@ export {
   LoadingOverlay,
   Screen,
   RoutingWrapper,
+  ProgressBar,
 };
