+++ feature-utils/poly-look/src/react-components/progressBar.css
@@ -6,7 +6,6 @@
     display: flex;
     flex-direction: row;
     box-sizing: border-box;
-    padding: 0 16px;
     margin: auto;
 }
 
@@ -48,7 +47,9 @@
     border-radius: 50%;
     font-size: 16px;
     font-weight: 500;
-    text-align: center;
+    display: flex;
+    justify-content: center;
+    align-items: center;
 }
 
 .progress-bar .section .request-number {
