--- features/google/src/components/progressBar/progressBar.css
@@ -1,68 +1,0 @@
-.import-view .progress-bar {
-    width: 100%;
-    max-width: var(--max-width);
-    height: 48px;
-    flex: 0 0 auto;
-    display: flex;
-    flex-direction: row;
-    box-sizing: border-box;
-    padding: 0 16px;
-    margin: auto;
-}
-
-.import-view .progress-bar .section {
-    flex: 1 1 0px;
-    height: 100%;
-    margin-right: 1px;
-    display: flex;
-    align-items: center;
-}
-
-.progress-bar .section .line {
-    width: 100%;
-    height: 4px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-}
-
-.progress-bar .request-progress {
-    background-color: var(--color-indigo-300);
-}
-
-.progress-bar .download-progress {
-    background-color: var(--color-indigo-500);
-}
-
-.progress-bar .import-progress {
-    background-color: var(--color-indigo-700);
-}
-
-.progress-bar .explore-progress {
-    background-color: var(--color-indigo-950);
-}
-
-.progress-bar .section .number {
-    width: 24px;
-    height: 24px;
-    border-radius: 50%;
-    font-size: 16px;
-    font-weight: 500;
-    text-align: center;
-}
-
-.progress-bar .section .request-number {
-    background-color: var(--color-indigo-300);
-}
-
-.progress-bar .section .download-number {
-    background-color: var(--color-indigo-500);
-}
-
-.progress-bar .section .import-number {
-    background-color: var(--color-indigo-700);
-}
-
-.progress-bar .section .explore-number {
-    background-color: var(--color-indigo-950);
-}
